#5 Feat: Create the context struct for the swap-tokens program and resolved the build errors in this solana contract...

```yaml

error: init_if_needed requires that anchor-lang be imported with the init-if-needed cargo feature enabled. Carefully read the init_if_needed docs before using this feature to make sure you know how to protect yourself against re-initialization attacks.
  --> programs/swap/src/instructions/accept_offer.rs:40:9
   |
40 |         init_if_needed,
   |         ^^^^^^^^^^^^^^

error[E0433]: failed to resolve: use of undeclared crate or module `anchor_spl`
 --> programs/swap/src/instructions/accept_offer.rs:2:5
  |
2 | use anchor_spl::token::{TokenAccount, Mint, Token}; // ✅ include Token
  |     ^^^^^^^^^^ use of undeclared crate or module `anchor_spl`

error[E0433]: failed to resolve: use of undeclared crate or module `anchor_spl`
 --> programs/swap/src/instructions/accept_offer.rs:3:5
  |
3 | use anchor_spl::associated_token::AssociatedToken;
  |     ^^^^^^^^^^ use of undeclared crate or module `anchor_spl`

error[E0433]: failed to resolve: use of undeclared crate or module `anchor_spl`
 --> programs/swap/src/instructions/create_offer.rs:2:5
  |
2 | use anchor_spl::token::{TokenAccount, Mint, Token}; // ✅ include Token
  |     ^^^^^^^^^^ use of undeclared crate or module `anchor_spl`

error[E0433]: failed to resolve: use of undeclared crate or module `anchor_spl`
 --> programs/swap/src/instructions/create_offer.rs:3:5
  |
3 | use anchor_spl::associated_token::AssociatedToken;
  |     ^^^^^^^^^^ use of undeclared crate or module `anchor_spl`

error[E0433]: failed to resolve: could not find `anchor_spl` in the list of imported crates
 --> programs/swap/src/instructions/create_offer.rs:8:10
  |
8 | #[derive(Accounts)]
  |          ^^^^^^^^ could not find `anchor_spl` in the list of imported crates
  |
  = note: this error originates in the derive macro `Accounts` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `custom-heap`
  --> programs/swap/src/lib.rs:16:1
   |
16 | #[program]
   | ^^^^^^^^^^
   |
   = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
   = help: consider adding `custom-heap` as a feature in `Cargo.toml`
   = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
   = note: `#[warn(unexpected_cfgs)]` on by default
   = note: this warning originates in the macro `$crate::custom_heap_default` which comes from the expansion of the attribute macro `program` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `custom-panic`
  --> programs/swap/src/lib.rs:16:1
   |
16 | #[program]
   | ^^^^^^^^^^
   |
   = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
   = help: consider adding `custom-panic` as a feature in `Cargo.toml`
   = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
   = note: this warning originates in the macro `$crate::custom_panic_default` which comes from the expansion of the attribute macro `program` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `anchor-debug`
 --> programs/swap/src/instructions/cancel_offer.rs:3:10
  |
3 | #[derive(Accounts)]
  |          ^^^^^^^^
  |
  = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
  = help: consider adding `anchor-debug` as a feature in `Cargo.toml`
  = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
  = note: this warning originates in the derive macro `Accounts` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `anchor-debug`
 --> programs/swap/src/instructions/create_offer.rs:8:10
  |
8 | #[derive(Accounts)]
  |          ^^^^^^^^
  |
  = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
  = help: consider adding `anchor-debug` as a feature in `Cargo.toml`
  = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
  = note: this warning originates in the derive macro `Accounts` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `anchor-debug`
 --> programs/swap/src/instructions/edit_offer.rs:4:10
  |
4 | #[derive(Accounts)]
  |          ^^^^^^^^
  |
  = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
  = help: consider adding `anchor-debug` as a feature in `Cargo.toml`
  = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
  = note: this warning originates in the derive macro `Accounts` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `anchor-debug`
  --> programs/swap/src/lib.rs:16:1
   |
16 | #[program]
   | ^^^^^^^^^^
   |
   = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
   = help: consider adding `anchor-debug` as a feature in `Cargo.toml`
   = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
   = note: this warning originates in the attribute macro `program` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: unexpected `cfg` condition value: `anchor-debug`
  --> programs/swap/src/lib.rs:16:1
   |
16 | #[program]
   | ^^^^^^^^^^
   |
   = note: expected values for `feature` are: `cpi`, `default`, `idl-build`, `no-entrypoint`, `no-idl`, `no-log-ix-name`, and `test-sbf`
   = help: consider adding `anchor-debug` as a feature in `Cargo.toml`
   = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
   = note: this warning originates in the derive macro `Accounts` (in Nightly builds, run with -Z macro-backtrace for more info)

error[E0277]: the trait bound `AcceptOffer<'_>: Bumps` is not satisfied
  --> programs/swap/src/instructions/accept_offer.rs:62:28
   |
62 | pub fn accept_offer(_ctx: &Context<AcceptOffer>) -> Result<()> {
   |                            ^^^^^^^^^^^^^^^^^^^^ the trait `Bumps` is not implemented for `AcceptOffer<'_>`
   |
   = help: the following other types implement trait `Bumps`:
             CreateNonceAccount<'info>
             CreateNonceAccountWithSeed<'info>
             __idl::IdlAccounts<'info>
             __idl::IdlCloseAccount<'info>
             __idl::IdlCreateAccounts<'info>
             __idl::IdlCreateBuffer<'info>
             __idl::IdlResizeAccount<'info>
             __idl::IdlSetBuffer<'info>
           and 14 others
note: required by a bound in `anchor_lang::context::Context`
  --> src/context.rs:24:1



```

+++ swap/programs/swap/src/constants.rs
@@ -1,4 +1,6 @@
 use anchor_lang::prelude::*;
 
+//3 | #[constant]
+//  | ^^^^^^^^^^^ function or associated item not found in `usize`
 #[constant]
-pub const ANCHOR_DISCRIMINATOR: usize = 8;
+pub const ANCHOR_DISCRIMINATOR: u8 = 8;
